First pass at autostyle.
This adds styles automatically for new layers, with a checkbox
that is by default checked. It covers all geometry types that
I know about and that TileMill can indicate. If it encounters
another one, or an unknown type, it won't insert anything except
for possibly a blank space.

We need to decide upon prettier styles if we want them - this
uses basic Mapnik styles. Also, it inserts at the end of the first
stylesheet. Inserting at the cursor won't work, because there's
too much possibility it could be inside of a rule. Inserting in
the active stylesheet would be better.