clarify where the shapefiles need to go
Small doc update to clarify where the shape files should go.
The documentation states that it should be put in the shp directory in the osm-bright folder but what is meant is the mapbox-osm-bright-* folder (ie. the root project folder). Also the OSM Bright repo does not contain a shp folder so we update the doc to reflect this.