Fixes to scripts to make them more robust.
Fixed the directory in loadosm.sh to use the config.json root if available or to use the default if not. Also made it properly handle spaces in the directory path.
Removed use of UTILS_DIR in scripts where it was not needed and made it relative elsewhere.
Changed the loadosm.sh to put a copy of defaults.style into the osm data dir (if not there) and then to use that one in case you have customized it.
changed the name of runtilemill.sh to tilemill.sh and moved it to the root.